+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## 3.4.0 [unreleased]
 
+### Breaking Changes
+Changed type of `Duration.magnitude` from `int?` to `long?`.
+
+### Features
+1. [#282](https://github.com/influxdata/influxdb-client-csharp/pull/282): Add support for AggregateWindow function [LINQ]
+
 ## 3.3.0 [2022-02-04]
 
 ### Bug Fixes